feat(upload): decouple finalize from scanner availability

+++ services/api/config/scanner.php
@@ -2,5 +2,5 @@
 
 return [
     'base_url' => env('SCANNER_BASE_URL', 'http://scanner:3001'),
-    'timeout'  => env('SCANNER_TIMEOUT', 10), // seconds
+    'timeout'  => (int) env('SCANNER_TIMEOUT', 10),
 ];
